[equipped-item-search] move : 캐릭터 스텟 모델 클래스 패키지 변경
